pyproject.toml
[tool.poetry]
name = "mqtt2file"
version = "0.1.0"
description = ""
authors = ["René Kübler <rnkblr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
